work updates

+++ work/MTRIX/Collect_r_factor_test_results.py
@@ -16,53 +16,61 @@ def run(directory_path):
     probelm_files = []
     for file in files:
         d = open(os.path.join(directory_path, file), "r").readlines()
+        d = open(os.path.join(directory_path, file), "r").read().splitlines()
         file = file[4:]
-        # files with errors have more than one line
-        ld =  len(d)
         # check every line in d if it can be splited
         msg = []
-        for l in d:
-            l_parts = l.split('::')
+        l_parts = d[0].split('::')
+        file_ok = False
+        if len(d) == 1:
+            # Check if the log file starts with the file name
             if l_parts[0] == file:
                 # check if this is the line with the r values
-                if len(l_parts)>3:
-                    l_parts = [x.strip() for x in l_parts]
-                    # check if score is OK
-                    if l_parts[2]<'100':
-                        temp_data_files = [file,float(l_parts[1]),float(l_parts[2]),float(l_parts[3])]
-                        msg.append('::'.join(l_parts[4:]))
-                        #temp_data.append(float(min(l_parts[1:3])))
-                    else:
-                        # collect files with issues
-                        new_record = [file,'::'.join(l_parts[4:])]
-                        probelm_files.append(new_record)
+                # check if score is OK
+                if l_parts[2]<'100':
+                    # R value < 100 means that the file was processed correctly
+                    temp_data_files = [file,float(l_parts[1]),float(l_parts[2]),float(l_parts[3])]
+                    file_ok = True
+                else:
+                    # file with issues
+                    file_ok = False
+
             # add information on rotation matix problems
             elif l_parts[0].startswith('Rotation matrices are not proper!'):
                 msg.append('::Rotation matrices are not proper!')
-        temp_data_files.extend(msg)
-        data_files.append(temp_data_files)
+                file_ok = False
+            else:
+                msg.append('::File was not processed properly!')
+                file_ok = False
+        # add errors from first line in file
+        msg.append('::'.join(l_parts[4:]))
+        if file_ok:
+            temp_data_files.extend(msg)
+            data_files.append(temp_data_files)
+        else:
+            temp_data_files = [file,msg]
+            probelm_files.append(temp_data_files)
 
-
+    print 'total number of files to process: {}'.format(len(files))
     print 'number of files with good data line: {}'.format(len(data_files))
     print 'number of files with problems: {}'.format(len(probelm_files))
-    # plot results
-    #xend = len(data)+1
-    #x = range(1, xend)
-    #plb.plot(x,data,'o')
-    #plb.show()
-
     return data_files, probelm_files
 
 def add_to_data_files(data_files, probelm_files,write_files=False):
     osType = sys.platform
     if osType.startswith('win'):
-        directory_path = 'c:\Phenix\Dev\Work\work'
+        directory_path = 'c:\Phenix\Dev\Work\work\MTRIX\Data'
     else:
-        directory_path = '/net/cci-filer2/raid1/home/youval/Work/work'
+        directory_path = '/net/cci-filer2/raid1/home/youval/Work/work/MTRIX/Data'
     os.chdir(directory_path)
-    print os.getcwd()
+    print 'add data in {}'.format(os.getcwd())
     if write_files:
-        pickle.dump(data_files, open('Collect_tested_files','w'))
+        if os.path.isfile('Collect_tested_files'):
+            Collect_tested_files = pickle.load(open('Collect_tested_files',"r"))
+            Collect_tested_files.extend(data_files)
+            pickle.dump(Collect_tested_files, open('Collect_tested_files','w'))
+        else:
+            pickle.dump(data_files, open('Collect_tested_files','w'))
         pickle.dump(probelm_files, open('files_with_problems','w'))
 
 
@@ -71,13 +79,12 @@ if __name__=='__main__':
     # locate the directory containing the log files
     osType = sys.platform
     if osType.startswith('win'):
-        directory_path = 'c:\Phenix\Dev\Work\work\queue_job'
+        directory_path = r'c:\Phenix\Dev\Work\work\junk\queue_job_3'
     else:
-        directory_path = '/net/cci-filer2/raid1/home/youval/Work/work/queue_job'
+        directory_path = '/net/cci-filer2/raid1/home/youval/Work/work/junk/queue_job_3'
     print os.getcwd()
     # convert the path to python format
     directory_path = os.path.realpath(directory_path)
     data_files, probelm_files = run(directory_path)
-    print os.getcwd()
-    #os.chdir('c:\\Phenix\\Dev\\Work\\work')
+    #
     #add_to_data_files(data_files, probelm_files,write_files=True)
